Refactor host/terminal views and shared enums
Major refactor across app, backend, and shared code:

- Remove RequestNotif.vue and stop rendering it in HostingLobby.
- Update router: set home to Waiting HomeScreen and rename host route to 'HostingRoom'.
- Simplify HostView: streamline socket hookup and use serverToHostRemotes constants.
- Split TerminalRoom: create TerminalRoom_Legacy.vue with previous implementation and reduce current TerminalRoom to use StationStore (log station type).
- HostingLobby: add navigation to HostingRoom when game starts and remove notification component.
- Backend: add safety check when resolving internal station data in attemptGameStart to avoid missing station types.
- gameLoop: use shared enums constant for NEW_RECIPE when emitting to host.
- shared: add boiling/frying/dicing stationData entries and rename/standardize enums (NEW_RECIPE, GAME_STARTED).

These changes align event naming with shared enums, tidy UI view responsibilities, and add defensive checks to backend station handling.
