+++ app/src/views/Waiting/HostingLobby.vue
@@ -2,7 +2,6 @@
     <p> ~ HOSTING LOBBY ~ </p>
 
     <CodeArea />
-    <RequestNotif />
     <UniqueIdentifier />
     <div>Current station count: {{ numberOfStations }}</div>
     <div>Current player count: {{ numberOfPlayers }}</div>
@@ -14,13 +13,14 @@
 import { useNotificationStore } from '@/stores/NotificationStore'
 import { useSocketStore } from '@/stores/SocketStore'
 import { uniqueIdentifier } from '@shared/types'
+import { useRouter } from 'vue-router'
 import { ref } from 'vue'
 
 import CodeArea from '@/components/Waiting/CodeArea.vue'
-import RequestNotif from '@/components/Waiting/RequestNotif.vue'
 import DecisionButton from '@/components/Shared/DecisionButton.vue'
 import UniqueIdentifier from '@/components/Shared/UniqueIdentifier.vue'
 
+const router = useRouter()
 const socketStore = useSocketStore()
 const notificationStore = useNotificationStore()
 
@@ -30,6 +30,10 @@ const notificationHandlerKeys = notificationStore.HANDLER_KEYS
 
 const numberOfPlayers = ref(0)
 const numberOfStations = ref(0)
+
+function routeToHostRoom() {
+    router.push({name: 'HostingRoom'})
+}
 
 function listenForPlayersPendingJoin() {
     socketStore.attachEventListener(ServerToHostRemotes.clientPendingJoin, (identifier: uniqueIdentifier) => {
@@ -70,6 +74,8 @@ function listenForGameStarted() {
         removeListenersForPlayersPendingJoin() 
         removeListenersForPlayersJoined()
         removeGameStartedListener()
+
+        routeToHostRoom()
     })
 }
